vendor apache-commons
Removes the dependency on apache-commons-text (and apache-commons-lang, transitively) by vendoring only the methods used in runelite-client. Some minimal changes were made during the vendoring, such as collapsing some classes and reformatting, but no functional changes were made to the source material.

This does not remove the use of apache-commons in cache tools.